+++ internal/infra/database/redis/repository/get-rows-users.go
@@ -10,13 +10,11 @@ func (r *MesssageRepository) GetUsersRows() (float64, error) {
 	ctx := context.Background()
 
 	totalUsers, err := r.rdb.LLen(ctx, "users").Result()
-
 	if err != nil {
 		return 0, err
 	}
 
-	divi := float64(totalUsers / entity.PerPage)
+	divi := float64(totalUsers) / float64(entity.PerPage)
 
 	return divi, nil
-
 }
